\name{rat2302rnentrezgfrmavecs}
\docType{data}
\alias{rat2302rnentrezgfrmavecs}
\title{Vectors used by fRMA for rat2302rnentrezg.}
\description{This data object was automatically created by the package frmaTools version 1.34.0.}
\usage{data(rat2302rnentrezgfrmavecs)}
\format{A list with 6 elements.
  \tabular{ll}{
    normVec \tab normalization vector \cr
    probeVec \tab probe effect vector \cr
    probeVarWithin \tab within batch probe variance \cr
    probeVarBetween \tab between batch probe variance \cr
    probesetSD \tab within probeset standard deviation \cr
    medianSE \tab median standard errors \cr
}}

\keyword{datasets}
\examples{
 data(rat2302rnentrezgfrmavecs) 
 str(rat2302rnentrezgfrmavecs)
}
